fix(*): fix ISR finally

+++ apps/web/src/items/mutations/deleteItem.ts
@@ -1,8 +1,4 @@
 import { resolver } from "@blitzjs/rpc"
-import { enforceSuperAdminIfNotCurrentOrganization } from "src/auth/helpers/enforceSuperAdminIfNoCurrentOrganization"
-import { setDefaultOrganizationId } from "src/auth/helpers/setDefaultOrganizationId"
-import { setDefaultVenue } from "src/auth/helpers/setDefaultVenue"
-import { revalidateVenue } from "src/core/helpers/revalidation"
 import db from "db"
 import { z } from "zod"
 
@@ -13,13 +9,9 @@ const DeleteItem = z.object({
 export default resolver.pipe(
   resolver.zod(DeleteItem),
   resolver.authorize(),
-  setDefaultOrganizationId,
-  enforceSuperAdminIfNotCurrentOrganization,
-  setDefaultVenue,
-  async ({ id, venue }) => {
+  async ({ id }, ctx) => {
     // TODO: in multi-tenant app, you must add validation to ensure correct tenant
-    const item = await db.item.deleteMany({ where: { id } })
-    await revalidateVenue(venue.identifier)()
+    const item = await db.item.deleteMany({ where: { id, organizationId: ctx.session.orgId } })
     return item
   }
 )
